requirements.txt
Flask==2.2.5
requests
google-cloud-storage
